requirements.txt
click==7.1.2
dlib
face-recognition==1.3.0
face-recognition-models==0.3.0
imutils==0.5.4
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
aiofiles
anyio
click
colorama
fastapi
greenlet
h11
idna
pydantic
python-multipart
sniffio
SQLAlchemy
starlette
typing_extensions
uvicorn
